Move HtmlSchema tests to their own project
