+++ dotNetRDF.sln
@@ -80,6 +80,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "dotNetRdf.Data.DataTables.T
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "dotNetRdf.Query.FullText.Tests", "Testing\dotNetRdf.Query.FullText.Tests\dotNetRdf.Query.FullText.Tests.csproj", "{7467A172-3DA1-44C2-9326-7883D8E550F6}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "dotNetRdf.Sparql.Tests", "Testing\dotNetRdf.Sparql.Tests\dotNetRdf.Sparql.Tests.csproj", "{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "dotNetRdf.HtmlSchema.Tests", "Testing\dotNetRdf.HtmlSchema.Tests\dotNetRdf.HtmlSchema.Tests.csproj", "{77F45239-2333-490D-BFDA-B7AA289EF83D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -474,6 +478,30 @@ Global
 		{7467A172-3DA1-44C2-9326-7883D8E550F6}.Release|x64.Build.0 = Release|Any CPU
 		{7467A172-3DA1-44C2-9326-7883D8E550F6}.Release|x86.ActiveCfg = Release|Any CPU
 		{7467A172-3DA1-44C2-9326-7883D8E550F6}.Release|x86.Build.0 = Release|Any CPU
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}.Debug|x64.Build.0 = Debug|Any CPU
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}.Debug|x86.Build.0 = Debug|Any CPU
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}.Release|x64.ActiveCfg = Release|Any CPU
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}.Release|x64.Build.0 = Release|Any CPU
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}.Release|x86.ActiveCfg = Release|Any CPU
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD}.Release|x86.Build.0 = Release|Any CPU
+		{77F45239-2333-490D-BFDA-B7AA289EF83D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{77F45239-2333-490D-BFDA-B7AA289EF83D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{77F45239-2333-490D-BFDA-B7AA289EF83D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{77F45239-2333-490D-BFDA-B7AA289EF83D}.Debug|x64.Build.0 = Debug|Any CPU
+		{77F45239-2333-490D-BFDA-B7AA289EF83D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{77F45239-2333-490D-BFDA-B7AA289EF83D}.Debug|x86.Build.0 = Debug|Any CPU
+		{77F45239-2333-490D-BFDA-B7AA289EF83D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{77F45239-2333-490D-BFDA-B7AA289EF83D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{77F45239-2333-490D-BFDA-B7AA289EF83D}.Release|x64.ActiveCfg = Release|Any CPU
+		{77F45239-2333-490D-BFDA-B7AA289EF83D}.Release|x64.Build.0 = Release|Any CPU
+		{77F45239-2333-490D-BFDA-B7AA289EF83D}.Release|x86.ActiveCfg = Release|Any CPU
+		{77F45239-2333-490D-BFDA-B7AA289EF83D}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -511,6 +539,8 @@ Global
 		{B5D33B8E-0F35-43D0-B649-2F0822FE5D78} = {3B3CCD52-CF1B-492C-A5F8-FBB626C5039C}
 		{4387CBC4-EDFD-4D48-AEB4-C340609B8F2D} = {F0F45D49-20AC-4407-8618-8C3650E34B97}
 		{7467A172-3DA1-44C2-9326-7883D8E550F6} = {F0F45D49-20AC-4407-8618-8C3650E34B97}
+		{953D71C0-80A3-4CB0-9A3B-1386834F2FBD} = {F0F45D49-20AC-4407-8618-8C3650E34B97}
+		{77F45239-2333-490D-BFDA-B7AA289EF83D} = {F0F45D49-20AC-4407-8618-8C3650E34B97}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {801E5D6F-1615-444B-BDE4-8A6A30FE0520}
